test+fix: cover the _v1 anthropic branch; align count_tokens fallback
Non-blocking review follow-ups (#4, #6 + the untested-anthropic-branch
inline comment):

- #6: replace the broken `except (AttributeError, Exception)` in
  count_tokens (rag_v1, reasoning_v1) with the three-clause chain the
  other forks already use (AttributeError / network / unexpected, each
  logged) so a transient tokenizer error no longer silently swaps in a
  different tiktoken encoding with no log line. Fix the "enconding" typo.
- #4: annotate `max_tokens` as Optional[int] (not float) across all four
  forks; anthropic 0.109.1 messages.create requires int.
- Tests: add the anthropic-branch coverage the suites were missing
  (they mock LLMClientManager wholesale). For each fork: the
  LLMClient.completions anthropic mapping (system extraction,
  content[0].text, input_tokens/output_tokens -> prompt/completion_tokens)
  and with_key_rotation's anthropic.RateLimitError handling (rotate the
  anthropic pool only; exhaustion re-raises). For the three napthaai forks
  add the with-client count_tokens path. The rag/url_cot forks also assert
  the caller's messages list is not mutated (regresses the deepcopy fix).

Re-locked: prediction_request_v1, prediction_request_rag_v1,
prediction_request_reasoning_v1, prediction_url_cot_v1, agent, service.

+++ packages/napthaai/customs/prediction_url_cot_v1/prediction_url_cot_v1.py
@@ -196,7 +196,7 @@ class LLMClient:
         top_p: Optional[float] = None,
         n: Optional[int] = None,
         stop: Any = None,
-        max_tokens: Optional[float] = None,
+        max_tokens: Optional[int] = None,
     ) -> Optional[LLMResponse]:
         """Generate a completion from the specified LLM provider using the given model and messages."""
         if self.llm_provider == "anthropic":
